Connection check

+++ client-android/app/src/main/java/com/dn/todo/ui/screen/TaskListScreen.kt
@@ -7,25 +7,31 @@ import androidx.compose.foundation.layout.RowScope
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.lazy.LazyColumn
+import androidx.compose.foundation.rememberScrollState
 import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.foundation.verticalScroll
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.FloatingActionButton
 import androidx.compose.material3.Icon
 import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.SnackbarHostState
 import androidx.compose.material3.SwipeToDismissBox
 import androidx.compose.material3.SwipeToDismissBoxValue
+import androidx.compose.material3.Text
 import androidx.compose.material3.pulltorefresh.PullToRefreshContainer
 import androidx.compose.material3.pulltorefresh.rememberPullToRefreshState
 import androidx.compose.material3.rememberSwipeToDismissBoxState
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.DisposableEffect
 import androidx.compose.runtime.LaunchedEffect
+import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.input.nestedscroll.nestedScroll
 import androidx.compose.ui.platform.LocalLifecycleOwner
 import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
 import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.lifecycle.Lifecycle
@@ -34,21 +40,31 @@ import androidx.paging.LoadState
 import androidx.paging.compose.collectAsLazyPagingItems
 import com.dn.todo.ui.R
 import com.dn.todo.ui.component.TaskCard
+import com.dn.todo.ui.theme.AppTypography
 import com.dn.todo.ui.viewmodel.TaskListViewModel
+import kotlinx.coroutines.launch
 
-private val TAG = "TaskListScreen"
+private const val TAG = "TaskListScreen"
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
 fun TaskListScreen(
     viewModel: TaskListViewModel = hiltViewModel(),
+    snackbarState: SnackbarHostState,
     onEditTask: (taskId: Long) -> Unit
 ) {
+    val scope = rememberCoroutineScope()
     val tasks = viewModel.tasks.collectAsLazyPagingItems()
     val lifecycleOwner = LocalLifecycleOwner.current
     val pullToRefresh = rememberPullToRefreshState() // ui indicator state
 
+    /* Resources */
+    val noConnectionMessage = stringResource(R.string.no_connection)
+
     /* UI Logic */
+
+    fun showSnackbarNoConnection() =
+        scope.launch { snackbarState.showSnackbar(noConnectionMessage) }
 
     // start loading when "pulled to refresh"
     LaunchedEffect(pullToRefresh.isRefreshing) {
@@ -58,12 +74,15 @@ fun TaskListScreen(
         }
     }
 
-    // hide refresh indicator when loaded
-    LaunchedEffect(tasks.loadState.refresh) {
+    LaunchedEffect(tasks.loadState) {
+        // hide refresh indicator when loaded
         if (tasks.loadState.refresh !is LoadState.Loading) {
             Log.d(TAG, "Hide pull-to-refresh indicator")
             pullToRefresh.endRefresh()
         }
+        // notify if no connection
+        if (tasks.loadState.hasError)
+            showSnackbarNoConnection()
     }
 
     // subscribe to data-driven events
@@ -72,6 +91,7 @@ fun TaskListScreen(
             when (it) {
                 is TaskListViewModel.DataDrivenEvent.RefreshTaskList -> tasks.refresh()
                 is TaskListViewModel.DataDrivenEvent.NavigateToTaskEdit -> onEditTask(it.taskId)
+                is TaskListViewModel.DataDrivenEvent.NotifyNoConnection -> showSnackbarNoConnection()
             }
         }
     }
@@ -93,28 +113,43 @@ fun TaskListScreen(
             .fillMaxSize()
             .nestedScroll(pullToRefresh.nestedScrollConnection)
     ) {
-        LazyColumn {
-            items(
-                count = tasks.itemCount,
-                key = { tasks[it]!!.id }
+        if (tasks.itemCount == 0) // pull-to-refresh indicator requires scrollable placeholder
+            Box(
+                Modifier.fillMaxSize()
+                    .verticalScroll(rememberScrollState())
             ) {
-                val item = tasks[it]!!
+                Text(
+                    modifier = Modifier.align(Alignment.Center),
+                    text = stringResource(R.string.no_tasks),
+                    style = AppTypography.titleLarge
+                )
+            }
+        else
+            LazyColumn {
+                items(
+                    count = tasks.itemCount,
+                    key = { tasks[it]!!.id }
+                ) {
+                    val item = tasks[it]!!
 
-                Box(Modifier.padding(horizontal = 16.dp, vertical = 8.dp)) {
-                    SwipeToRemoveCard(onRemoved = { viewModel.deleteTask(item.id) }) {
-                        TaskCard(
-                            item.title, item.description, item.isCompleted,
-                            onCheckedChange = { isCompleted -> viewModel.markTask(item.id, isCompleted) },
-                            onClick = { viewModel.editTask(item.id) }
-                        )
+                    Box(Modifier.padding(horizontal = 16.dp, vertical = 8.dp)) {
+                        SwipeToRemoveCard(onRemoved = { viewModel.deleteTask(item.id) }) {
+                            TaskCard(
+                                item.title, item.description, item.isCompleted,
+                                onCheckedChange = { isCompleted ->
+                                    viewModel.markTask(item.id, isCompleted)
+                                },
+                                onClick = { viewModel.editTask(item.id) }
+                            )
+                        }
                     }
                 }
             }
-        }
+
         FloatingActionButton(
             modifier = Modifier
                 .align(Alignment.BottomEnd)
-                .padding(16.dp),
+                .padding(24.dp),
             onClick = { viewModel.createTask() }
         ) {
             Icon(painter = painterResource(id = R.drawable.ic_add), contentDescription = null)
